requirements.txt
django-admin-rangefilter==0.5.3
django-autocomplete-light==3.5.0
django-debug-toolbar==2.1
django-extensions==2.2.5
django-import-export==2.0
django-json-widget==0.2.0
Django==2.2.9
djangorestframework==3.11.0
drf-yasg==1.17.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
sentry-sdk==0.13.5
pandas==1.1.5